+++ app/config/config.neon
@@ -59,6 +59,7 @@ services:
 	spravceOblasti: App\Model\SpravceOblasti
 	cestneClenstviUzivatele: App\Model\CestneClenstviUzivatele
 	log: App\Model\Log
+	pdfGenerator: App\Model\PdfGenerator
 	povoleneSMTP: App\Model\PovoleneSMTP
 	accountActivation: App\Model\AccountActivation
 	logTableFactory: App\Components\LogTableFactory
