+++ app/model/Uzivatel.php
@@ -275,21 +275,4 @@ ORDER BY t1.id LIMIT 1')->fetchField();
         }
         return null;
     }
-
-    public function mesicName($indate, $addmonth){
-        $date = new Nette\Utils\DateTime($indate);
-        $date->add(new \DateInterval('P'.$addmonth.'M'));
-        $datestr = $date->format('F');
-
-        $aj = array("January","February","March","April","May","June","July","August","September","October","November","December");
-        $cz = array("leden","únor","březen","duben","květen","červen","červenec","srpen","září","říjen","listopad","prosinec");
-        $datum = str_replace($aj, $cz, $datestr);
-        return $datum;
-    }
-
-    public function mesicDate($indate, $addmonth){
-        $date = new Nette\Utils\DateTime($indate);
-        $date->add(new \DateInterval('P'.$addmonth.'M'));
-        return $date->format('17.m.Y');
-    }
 }
